# Conflicts:
#	ITMartinFileSorter.Server/Program.cs

+++ ITMartinBudget.Application/Models/TransactionRule.cs
@@ -14,7 +14,7 @@ public class TransactionRule
 
     public BudgetGroup BudgetGroup { get; set; }
 
-    public int RecurringIntervalMonths { get; set; }
+    public decimal RecurringIntervalMonths { get; set; }
     public TransactionType TransactionType { get; set; }
     public int Priority { get; init; }
     public ComparingType ComparingType { get; set; }
